DescriptorValidator
