+++ com.sap.cloud.lm.sl.mta/src/main/java/com/sap/cloud/lm/sl/mta/handlers/v2/DescriptorValidator.java
@@ -9,39 +9,34 @@ import com.sap.cloud.lm.sl.mta.validators.v2.DeploymentDescriptorValidator;
 import com.sap.cloud.lm.sl.mta.validators.v2.ExtensionDescriptorValidator;
 import com.sap.cloud.lm.sl.mta.validators.v2.MergedDescriptorValidator;
 
-public class DescriptorValidator extends com.sap.cloud.lm.sl.mta.handlers.v1.DescriptorValidator {
+public class DescriptorValidator {
+
+    protected final DescriptorHandler handler;
 
     public DescriptorValidator() {
-        super(new DescriptorHandler());
+        this(new DescriptorHandler());
     }
 
     public DescriptorValidator(DescriptorHandler handler) {
-        super(handler);
+        this.handler = handler;
     }
 
-    @Override
-    protected DeploymentDescriptorValidator getDeploymentDescriptorValidator(
-        com.sap.cloud.lm.sl.mta.model.v1.DeploymentDescriptor deploymentDescriptor,
-        com.sap.cloud.lm.sl.mta.model.v1.Platform platformType) {
-        return new DeploymentDescriptorValidator((DeploymentDescriptor) deploymentDescriptor, (Platform) platformType,
-            (DescriptorHandler) handler);
+    protected DeploymentDescriptorValidator getDeploymentDescriptorValidator(DeploymentDescriptor deploymentDescriptor,
+        Platform platformType) {
+        return new DeploymentDescriptorValidator(deploymentDescriptor, platformType, handler);
     }
 
-    @Override
-    protected ExtensionDescriptorValidator getExtensionDescriptorValidator(
-        com.sap.cloud.lm.sl.mta.model.v1.ExtensionDescriptor extensionDescriptor,
-        com.sap.cloud.lm.sl.mta.model.v1.DeploymentDescriptor deploymentDescriptor) {
-        return new ExtensionDescriptorValidator((ExtensionDescriptor) extensionDescriptor, (DeploymentDescriptor) deploymentDescriptor,
-            (DescriptorHandler) handler);
+    protected ExtensionDescriptorValidator getExtensionDescriptorValidator(ExtensionDescriptor extensionDescriptor,
+        DeploymentDescriptor deploymentDescriptor) {
+        return new ExtensionDescriptorValidator(extensionDescriptor, deploymentDescriptor, handler);
     }
 
-    @Override
-    protected MergedDescriptorValidator getMergedDescriptorValidator(com.sap.cloud.lm.sl.mta.model.v1.DeploymentDescriptor mergedDescriptor,
+
+    protected MergedDescriptorValidator getMergedDescriptorValidator(DeploymentDescriptor mergedDescriptor,
         DescriptorValidationRules validationRules) {
-        return new MergedDescriptorValidator((DeploymentDescriptor) mergedDescriptor, validationRules, (DescriptorHandler) handler);
+        return new MergedDescriptorValidator(mergedDescriptor, validationRules, handler);
     }
 
-    @Override
     protected DescriptorValidationRules getDefaultDescriptorValidationRules() {
         return new DefaultDescriptorValidationRules();
     }
